# #104 improvement: EntitySet.InterestingEntities now uses HashSet