As requested in #105 , EntitySet.InterestingEntities now uses HashSet. Benchmarks on my machine:

current master:
``` ini

BenchmarkDotNet=v0.10.12, OS=Windows 10 Redstone 3 [1709, Fall Creators Update] (10.0.16299.192)
Intel Core i3 CPU M 370 2.40GHz, 1 CPU, 4 logical cores and 2 physical cores
Frequency=2337902 Hz, Resolution=427.7339 ns, Timer=TSC
  [Host]   : .NET Framework 4.6.1 (CLR 4.0.30319.42000), 32bit LegacyJIT-v4.7.2600.0  [AttachedDebugger]
  ShortRun : .NET Framework 4.6.1 (CLR 4.0.30319.42000), 32bit LegacyJIT-v4.7.2600.0

Job=ShortRun  LaunchCount=1  TargetCount=3  
WarmupCount=3  

```
|          Method | NumEntities |           Mean |        Error |        StdDev |      Gen 0 |     Gen 1 |     Gen 2 |   Allocated |
|---------------- |------------ |---------------:|-------------:|--------------:|-----------:|----------:|----------:|------------:|
|            **None** |          **10** |       **523.4 us** |     **323.7 us** |     **18.290 us** |    **57.6172** |         **-** |         **-** |    **88.77 KB** |
|             Add |          10 |       788.1 us |     308.6 us |     17.435 us |    67.3828 |         - |         - |   104.36 KB |
|          Attach |          10 |       901.6 us |     420.7 us |     23.768 us |    80.0781 |         - |         - |   123.87 KB |
|    AddAndModify |          10 |     1,145.0 us |     790.1 us |     44.641 us |   101.5625 |         - |         - |   156.19 KB |
|    AddAndRemove |          10 |       821.9 us |     196.1 us |     11.082 us |    69.3359 |         - |         - |    106.9 KB |
|    AddAndDetach |          10 |       897.1 us |     391.2 us |     22.103 us |    82.0313 |         - |         - |    127.8 KB |
| AttachAndModify |          10 |     1,264.9 us |     925.3 us |     52.281 us |   111.3281 |         - |         - |   173.54 KB |
| AttachAndRemove |          10 |       994.2 us |     698.8 us |     39.485 us |    85.9375 |         - |         - |   132.54 KB |
| AttachAndDetach |          10 |       916.3 us |     137.8 us |      7.788 us |    81.0547 |         - |         - |   125.55 KB |
|            **None** |         **500** |    **29,733.3 us** |   **7,883.6 us** |    **445.439 us** |   **750.0000** |  **250.0000** |   **62.5000** |  **4403.28 KB** |
|             Add |         500 |    49,506.9 us |  31,634.4 us |  1,787.402 us |   875.0000 |  250.0000 |   62.5000 |  5171.44 KB |
|          Attach |         500 |    50,969.9 us |  30,525.2 us |  1,724.730 us |  1062.5000 |  375.0000 |  125.0000 |  6157.63 KB |
|    AddAndModify |         500 |    69,058.6 us |  13,394.5 us |    756.814 us |  1312.5000 |  312.5000 |  125.0000 |  7761.46 KB |
|    AddAndRemove |         500 |    52,845.3 us |   6,909.4 us |    390.394 us |   937.5000 |  250.0000 |   62.5000 |  5294.44 KB |
|    AddAndDetach |         500 |    56,306.4 us |  17,965.7 us |  1,015.093 us |  1062.5000 |  312.5000 |  125.0000 |   6339.5 KB |
| AttachAndModify |         500 |    74,158.9 us |  48,951.5 us |  2,765.848 us |  1437.5000 |  375.0000 |  125.0000 |  8632.66 KB |
| AttachAndRemove |         500 |    54,368.3 us |  41,144.7 us |  2,324.750 us |  1125.0000 |  312.5000 |  125.0000 |  6581.63 KB |
| AttachAndDetach |         500 |    53,371.2 us |  33,895.5 us |  1,915.159 us |  1062.5000 |  375.0000 |  125.0000 |  6239.62 KB |
|            **None** |        **5000** |   **336,174.5 us** | **196,147.9 us** | **11,082.719 us** |  **8187.5000** | **3750.0000** |  **937.5000** | **44090.49 KB** |
|             Add |        5000 | 1,051,342.1 us |  95,779.0 us |  5,411.694 us |  9750.0000 | 3125.0000 | 1250.0000 | 51826.22 KB |
|          Attach |        5000 |   559,888.6 us |  62,158.9 us |  3,512.094 us | 11125.0000 | 2750.0000 | 1062.5000 | 61642.68 KB |
|    AddAndModify |        5000 | 1,271,069.8 us | 124,063.2 us |  7,009.804 us | 14250.0000 | 3625.0000 | 1562.5000 | 77745.69 KB |
|    AddAndRemove |        5000 | 1,283,819.0 us | 219,889.6 us | 12,424.170 us |  9937.5000 | 3312.5000 | 1250.0000 | 53053.31 KB |
|    AddAndDetach |        5000 | 1,135,328.1 us |  68,841.3 us |  3,889.662 us | 11187.5000 | 2562.5000 |  812.5000 | 63496.67 KB |
| AttachAndModify |        5000 | 1,098,044.7 us | 294,381.5 us | 16,633.099 us | 15500.0000 | 3750.0000 | 1375.0000 | 86422.97 KB |
| AttachAndRemove |        5000 | 1,117,901.0 us |  42,999.6 us |  2,429.560 us | 11937.5000 | 2750.0000 |  937.5000 | 65900.79 KB |
| AttachAndDetach |        5000 |   584,212.0 us |  96,938.0 us |  5,477.178 us | 11187.5000 | 2562.5000 | 1000.0000 | 62459.97 KB |

This PR:
``` ini

BenchmarkDotNet=v0.10.12, OS=Windows 10 Redstone 3 [1709, Fall Creators Update] (10.0.16299.192)
Intel Core i3 CPU M 370 2.40GHz, 1 CPU, 4 logical cores and 2 physical cores
Frequency=2337902 Hz, Resolution=427.7339 ns, Timer=TSC
  [Host]   : .NET Framework 4.6.1 (CLR 4.0.30319.42000), 32bit LegacyJIT-v4.7.2600.0
  ShortRun : .NET Framework 4.6.1 (CLR 4.0.30319.42000), 32bit LegacyJIT-v4.7.2600.0

Job=ShortRun  LaunchCount=1  TargetCount=3  
WarmupCount=3  

```
|          Method | NumEntities |         Mean |        Error |       StdDev |      Gen 0 |     Gen 1 |     Gen 2 |   Allocated | Mean change %  |  Allocated change %  |
|---------------- |------------ |-------------:|-------------:|-------------:|-----------:|----------:|----------:|------------:|------------:|------------:|
|            **None** |          **10** |     **434.0 us** |     **31.22 us** |     **1.764 us** |    **57.6172** |         **-** |         **-** |    **88.81 KB** |  -17.08  |  0.05  |
  -17.08  ||             Add |          10 |     669.9 us |    209.67 us |    11.847 us |    67.3828 |         - |         - |   104.54 KB |  -15.00  |  0.17  |
|          Attach |          10 |     731.6 us |    336.47 us |    19.011 us |    80.0781 |         - |         - |   123.91 KB |  -18.86  |  0.03  |
|    AddAndModify |          10 |     919.5 us |    164.00 us |     9.266 us |   101.5625 |         - |         - |   156.38 KB |  -19.69  |  0.12  |
|    AddAndRemove |          10 |     678.8 us |    382.17 us |    21.593 us |    69.3359 |         - |         - |   107.07 KB |  -17.41  |  0.16  |
|    AddAndDetach |          10 |     737.0 us |    254.74 us |    14.393 us |    82.0313 |         - |         - |   127.97 KB |  -17.85  |  0.13  |
| AttachAndModify |          10 |   1,045.9 us |    381.55 us |    21.558 us |   111.3281 |         - |         - |   173.72 KB |  -17.31  |  0.10  |
| AttachAndRemove |          10 |     753.5 us |    128.13 us |     7.240 us |    85.9375 |         - |         - |   132.56 KB |  -24.21  |  0.02  |
| AttachAndDetach |          10 |     765.4 us |    155.39 us |     8.780 us |    81.0547 |         - |         - |   125.59 KB |  -16.47  |  0.03  |
|            **None** |         **500** |  **25,577.6 us** | **10,440.34 us** |   **589.899 us** |   **750.0000** |  **281.2500** |   **93.7500** |  **4403.43 KB** |  -13.98  |  0.00  |
|             Add |         500 |  41,127.9 us |  4,265.26 us |   240.995 us |   875.0000 |  250.0000 |   62.5000 |  5186.18 KB |  -16.92  |  0.29  |
|          Attach |         500 |  43,531.3 us | 34,284.00 us | 1,937.109 us |  1062.5000 |  375.0000 |  125.0000 |  6157.64 KB |  -14.59  |  0.00  |
|    AddAndModify |         500 |  57,635.9 us |  4,929.89 us |   278.548 us |  1312.5000 |  375.0000 |  125.0000 |  7776.44 KB |  -16.54  |  0.19  |
|    AddAndRemove |         500 |  41,638.7 us | 32,956.26 us | 1,862.090 us |   937.5000 |  250.0000 |   62.5000 |  5309.42 KB |  -21.21  |  0.28  |
|    AddAndDetach |         500 |  44,380.3 us | 34,654.27 us | 1,958.030 us |  1062.5000 |  312.5000 |  125.0000 |  6354.43 KB |  -21.18  |  0.24  |
| AttachAndModify |         500 |  58,924.5 us |  7,142.79 us |   403.581 us |  1437.5000 |  375.0000 |  125.0000 |  8647.33 KB |  -20.54  |  0.17  |
| AttachAndRemove |         500 |  43,957.7 us | 13,757.85 us |   777.344 us |  1125.0000 |  312.5000 |  125.0000 |  6588.81 KB |  -19.15  |  0.11  |
| AttachAndDetach |         500 |  43,584.0 us | 23,425.75 us | 1,323.598 us |  1062.5000 |  375.0000 |  125.0000 |  6239.65 KB |  -18.34  |  0.00  |
|            **None** |        **5000** | **282,600.7 us** | **73,862.69 us** | **4,173.379 us** |  **8187.5000** | **3750.0000** |  **937.5000** | **44090.49 KB** |  -15.94  |  0.00  |
|             Add |        5000 | 680,171.6 us | 45,921.55 us | 2,594.653 us |  9500.0000 | 2375.0000 | 1000.0000 | 51926.84 KB |  -35.30  |  0.19  |
|          Attach |        5000 | 475,084.4 us | 19,079.72 us | 1,078.039 us | 11125.0000 | 2750.0000 | 1062.5000 | 61641.86 KB |  -15.15  |  0.00  |
|    AddAndModify |        5000 | 868,108.4 us | 63,465.13 us | 3,585.897 us | 14312.5000 | 4000.0000 | 1625.0000 | 77846.51 KB |  -31.70  |  0.13  |
|    AddAndRemove |        5000 | 698,019.7 us | 32,491.61 us | 1,835.836 us |  9812.5000 | 3062.5000 | 1125.0000 | 53166.58 KB |  -45.63  |  0.21  |
|    AddAndDetach |        5000 | 741,620.2 us | 21,668.00 us | 1,224.282 us | 11937.5000 | 3625.0000 | 1562.5000 | 63628.12 KB |  -34.68  |  0.21  |
| AttachAndModify |        5000 | 667,719.9 us | 70,249.91 us | 3,969.250 us | 15500.0000 | 3687.5000 | 1375.0000 | 86540.97 KB |  -39.19  |  0.14  |
| AttachAndRemove |        5000 | 513,896.9 us | 63,658.55 us | 3,596.826 us | 11937.5000 | 2687.5000 |  937.5000 | 65928.85 KB |  -54.03  |  0.04  |
| AttachAndDetach |        5000 | 492,360.6 us | 37,795.47 us | 2,135.514 us | 11187.5000 | 2562.5000 | 1000.0000 |  62460.4 KB |  -15.72  |  0.00  |